Clean TypeScript OAuth boundary
